requirements.txt
# === Core ===
fastapi
uvicorn
streamlit
requests
python-dotenv
python-multipart

# === Vector DB & Embeddings ===
qdrant-client
langchain
langchain-community
langchain-qdrant
langchain-core
langchain-text-splitters
sentence-transformers
tiktoken

# === LLM ===
groq

# === Utilities ===
numpy
pandas
transformers
docx2txt
pypdf
pydantic
